Modularize routes.ts wave 2: extract projects + diagnostics
server/routes-modules/projects.ts  (131 lines)
  All /api/projects/* endpoints + /api/conversations/:id/project
  assignment. registerProjectRoutes(app).

server/routes-modules/diagnostics.ts  (175 lines)
  /api/system/runtime, /api/admin/system-status,
  /api/admin/provider-diagnostics. Takes an isDatabaseHealthy
  getter callback so it stays decoupled from routes.ts's
  module-level mutable.

routes.ts: 1,697 → 1,454 (-243 lines, -14%).
Combined with wave 1 (-404), total reduction so far is -647 lines
(2,101 → 1,454, -31%).
